Add comment to .env (#309)
Checking in a `.env` file into git is not best practice. We have historically included this file in this template project to make our getting started guide easier for beginners to consume and currently we'd like to continue that ease of use. From discussions in #296 a reasonable alternative to removing the file would be to add a comment in the `.env` file (as this PR does) which explains why the file is in git and how to remove it.

@fixes #47